backend/requirements.txt
# Crypto Trading Bot - Backend Requirements
# Flask and extensions
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
# Data processing
numpy==1.24.3
pandas==2.0.3
# Cryptography and security
pyjwt==2.8.0
cryptography==41.0.3
python-dotenv==1.0.0
# API and networking
requests==2.31.0
aiohttp==3.8.5
# Utilities
python-dateutil==2.8.2